Move FAQ to before last page
