+++ Froststrap/UI/ViewModels/Onboarding/Page5ViewModel.cs
@@ -1,38 +1,85 @@
-﻿using System.Windows.Input;
+using System.Collections.ObjectModel;
+using System.Windows.Input;
 using CommunityToolkit.Mvvm.Input;
-using Froststrap.UI.Elements.Onboarding;
 
 namespace Froststrap.UI.ViewModels.Onboarding
 {
+    public class FaqItem : NotifyPropertyChangedViewModel
+    {
+        private bool _isExpanded;
+
+        public string Question { get; set; } = string.Empty;
+        public string Answer { get; set; } = string.Empty;
+
+        public bool IsExpanded
+        {
+            get => _isExpanded;
+            set
+            {
+                _isExpanded = value;
+                OnPropertyChanged(nameof(IsExpanded));
+            }
+        }
+
+        public ICommand ToggleCommand { get; }
+
+        public FaqItem()
+        {
+            ToggleCommand = new RelayCommand(ToggleExpanded);
+        }
+
+        private void ToggleExpanded()
+        {
+            IsExpanded = !IsExpanded;
+        }
+    }
+
     public class Page5ViewModel : NotifyPropertyChangedViewModel
     {
-        public ICommand LaunchRobloxCommand { get; }
-        public ICommand LaunchSettingsCommand { get; }
+        public ObservableCollection<FaqItem> FaqItems { get; } = [];
 
         public Page5ViewModel()
         {
-            LaunchRobloxCommand = new RelayCommand(LaunchRoblox);
-            LaunchSettingsCommand = new RelayCommand(LaunchSettings);
+            LoadFaqData();
         }
 
-        private void LaunchRoblox()
+        private void LoadFaqData()
         {
-            var mainWindow = MainWindow.Instance;
-            if (mainWindow != null)
+            FaqItems.Add(new FaqItem
             {
-                mainWindow.CloseAction = NextAction.LaunchRoblox;
-                mainWindow.Close();
-            }
-        }
+                Question = Strings.Menu_Onboarding_Page6_Q1,
+                Answer = Strings.Menu_Onboarding_Page6_A1
+            });
 
-        private void LaunchSettings()
-        {
-            var mainWindow = MainWindow.Instance;
-            if (mainWindow != null)
+            FaqItems.Add(new FaqItem
             {
-                mainWindow.CloseAction = NextAction.LaunchSettings;
-                mainWindow.Close();
-            }
+                Question = Strings.Menu_Onboarding_Page6_Q2,
+                Answer = Strings.Menu_Onboarding_Page6_A2
+            });
+
+            FaqItems.Add(new FaqItem
+            {
+                Question = Strings.Menu_Onboarding_Page6_Q3,
+                Answer = Strings.Menu_Onboarding_Page6_A3
+            });
+
+            FaqItems.Add(new FaqItem
+            {
+                Question = Strings.Menu_Onboarding_Page6_Q4,
+                Answer = Strings.Menu_Onboarding_Page6_A4
+            });
+
+            FaqItems.Add(new FaqItem
+            {
+                Question = Strings.Menu_Onboarding_Page6_Q5,
+                Answer = Strings.Menu_Onboarding_Page6_A5
+            });
+
+            FaqItems.Add(new FaqItem
+            {
+                Question = Strings.Menu_Onboarding_Page6_Q6,
+                Answer = Strings.Menu_Onboarding_Page6_A6
+            });
         }
     }
 }
